Remove uintptr_t, segcache tests segfaulting. Now compiles on linux

+++ src/SegCache.h
@@ -243,7 +243,7 @@ union SegCachePrefixArray
     void ** raw;
     SegCachePrefixArray * array;
     SegCachePrefixEntry ** prefixEntries;
-    uintptr_t * range;
+    uint32 * range;
 };
 
 class SegCache
